requirements.txt
openai
PyQt5
sounddevice
scipy
speechrecognition
pyperclip
python-dotenv
pytesseract
pillow